Chore: main 페이지에서 header 컴포넌트 분리
